Move Function code to separate file
