+++ snaketalk/__init__.py
@@ -1,6 +1,7 @@
 from snaketalk.bot import Bot
+from snaketalk.function import Function, listen_to
 from snaketalk.message import Message
-from snaketalk.plugins import ExamplePlugin, Function, Plugin, listen_to
+from snaketalk.plugins import ExamplePlugin, Plugin
 from snaketalk.scheduler import schedule
 from snaketalk.settings import Settings
