+++ snaketalk/plugins/__init__.py
@@ -1,5 +1,5 @@
-from snaketalk.plugins.base import Function, Plugin, listen_to
+from snaketalk.plugins.base import Plugin
 from snaketalk.plugins.example import ExamplePlugin
 from snaketalk.plugins.webhook_example import WebhookExample
 
-__all__ = ["Function", "Plugin", "listen_to", "ExamplePlugin", "WebhookExample"]
+__all__ = ["Plugin", "ExamplePlugin", "WebhookExample"]
